requirements.txt
certifi==2019.3.9
chardet==3.0.4
fonttools==3.40.0
htmlmin==0.1.12
idna==2.8
Jinja2==2.10.1
libsass==0.18.0
MarkupSafe==1.1.1
pyalsaaudio==0.8.4
requests==2.21.0
six==1.12.0
urllib3==1.24.2
websockets==7.0
